Revert LayoutInline::CollectLineBoxRects
This patch reverts |LayoutInline::CollectLineBoxRects| not to
use |NGInlineCursor|, by manually merging parts of r706815
<crrev.com/c/1862534> and r710694 <crrev.com/c/1888624>.

|NGInlineCursor| is rapidly changing that reverting is not
easy. This patch reverts not to use it for |getClientRects|
and a few other callers.

Testing more on culled inline box indicates that the phase 1
of LayoutNG regressed some cases from legacy, and
|NGInlineCursor| work regressed further. We might want to fix
when NGFragmentItem ships, but better not to change the
behavior until then.

Bug: 1036358
Change-Id: I47a60f5d3a505d4b43eba0d8b5adc64e02cbcd28