[css-pseudo] Fix WPT expectations for marker-font-variant-numeric-*
The marker-font-variant-numeric-default.html test is supposedly passing
in Chromium according to Chromium, but not according to wpt.fyi. The
reason is that in certain configurations, the test may look slightly
different than the reference due to font anti-aliasing.

The marker-font-variant-numeric-normal.html fails in all browsers due to
lack of support for nested ::marker selectors, but there are also some
differences in the parts that should actually match the reference.

This patch fixes both issues.

Bug: 457718
Change-Id: Id646b7810aed18441e1c7b37ad7bb1a37877be2f