Portals: Use WebContents::ClosePage to run unload handlers for portals.
On the way, refactor how the Portal <-> WebContentsImpl pointers are
maintained, as this work caused the raw WebContentsImpl* to dangling while
writing the main code of this CL, which made for a confusing bug. This
refactoring is mainly a mechanical gathering of the code responsible for
updating these into a small nested class.

Bug: 964481
Change-Id: I8a63e1a5531de4807569b35ea0f3fcb19baf343c